requirements.txt
Django
django-crispy-forms
crispy-bootstrap5
psycopg2
pillow
pandas